WIP refactor weight types/impl out of runtimes to dedup code
