+++ libs/types/src/lib.rs
@@ -26,4 +26,5 @@ pub mod orders;
 pub mod permissions;
 pub mod time;
 pub mod tokens;
+pub mod weights;
 pub mod xcm;
